Using a resolution of 96 for mesh.
As can be seen in the pictures, a smaller resolution causes
problems at the dendrite/dendritic shaft boundaries.

I also toyed around with altering the number of slices used
(an optional parameter to dolfin.Cone) but this didn't actually
matter. I was under the impression that since each shaft is
10% the size of the dendrite that we'd need 10 times as many
slices for the dendrite. However, it seems to equalize since
via CSG the geometries come together.