Moving back to using the full mesh.
It preserves orientation and faces already know if
they are on the exterior. Using the full mesh, the
number of exterior faces (14730) matches very closely
with that of the boundary mesh (14808).

The number of top points with the full mesh (188) also
matches closely the number of top points from the
boundary mesh (203). However, when using the full mesh
the normal vectors correctly point out since they
come from a tetrahedral cell which provides them with
the correct orientation where in using the boundary
mesh only 102/203 of the faces had an upward pointing
normal vector (0, 0, 1), the other 101/203 had a downward
pointing normal vector (0, 0, -1).