This is a Monte Carlo simulation of an enclosure-escape test (a Poisson process)
This EXITBOX.m file is a function I wrote to perform a Monte Carlo
simulation of an enclosure-escape test (a Poisson process). This can
later be used to compute the Poisson distribution for the new 3D mesh
simulation. This primary-function will also provide/plot the actual
probability distribution for spine escape latencies for spines of
various shapes. Understanding the distribution for escape latencies is
one of the primary goals, as it will tell us about how spines can
capture the most particles for the longest duration.